Remove differently

+++ Commands.cpp
@@ -84,7 +84,7 @@ void _removeBackgroundSign(char* cmd_line) {
 
 // TODO: Add your implementation for classes in Commands.h 
 
-SmallShell::SmallShell(): prompt("smash> "), last_working_dir("") {
+SmallShell::SmallShell(): prompt("smash> "), last_working_dir(""), smash_pid(getpid()), pid_in_fg(0) {
 
 }
 
@@ -250,7 +250,8 @@ ShowPidCommand::ShowPidCommand(const char *cmd_line) : BuiltInCommand(cmd_line) 
 }
 
 void ShowPidCommand::execute() {
-    cout << "smash pid is " << getpid() << endl;
+    SmallShell& smash = SmallShell::getInstance();
+    cout << "smash pid is " << smash.smash_pid << endl;
 }
 
 GetCurrDirCommand::GetCurrDirCommand(const char *cmd_line) : BuiltInCommand(cmd_line){
@@ -441,7 +442,8 @@ JobEntry * JobsList::getLastStoppedJob(int *jobId) {
 }
 
 void JobsList::removeFinishedJobs() {
-    for(int i=0 ;i<jobs.size();i++){
+    vector<int> to_be_removed;
+    for(int i=0 ;i<jobs.size(); i++){
         int status;
         int res = waitpid(jobs[i]->process_id, &status,WNOHANG | WUNTRACED | WCONTINUED);
         if (res==-1){
@@ -449,13 +451,11 @@ void JobsList::removeFinishedJobs() {
             return;
         }
         if(res > 0 && !jobs[i]->is_stopped && !WIFCONTINUED(status)){
-            JobEntry* temp = jobs[i];
-            // cout << "Removing job: " << temp->job_id << endl;
-            jobs.erase(jobs.begin()+i);
-            delete temp;
-        } else {
-            i++;
+            to_be_removed.push_back(jobs[i]->job_id);
         }
+    }
+    for (int i=0; i<to_be_removed.size(); i++){
+        removeJobById(to_be_removed[i]);
     }
 }
 
@@ -638,6 +638,7 @@ void ForegroundCommand::execute() {
               perror("smash error: kill failed");
               return;
           }
+          our_job->is_stopped = false;
       }
       int status;
       smash.pid_in_fg=our_job->process_id ;
@@ -645,6 +646,9 @@ void ForegroundCommand::execute() {
       if(waitpid(our_job->process_id,&status,WUNTRACED)==-1){
           perror("smash error: waitpid failed");
           return;
+      }
+      if (!our_job->is_stopped) {
+          smash.jobs_list.removeJobById(our_job->job_id);
       }
 
       smash.pid_in_fg = 0;
@@ -674,6 +678,7 @@ void ForegroundCommand::execute() {
                     perror("smash error: kill failed");
                     return;
                 }
+                our_job->is_stopped = false;
             }
             int status;
             smash.pid_in_fg = our_job->process_id;
@@ -681,6 +686,9 @@ void ForegroundCommand::execute() {
             if (waitpid(our_job->process_id, &status, WUNTRACED) == -1) {
                 perror("smash error: waitpid failed");
                 return;
+            }
+            if (!our_job->is_stopped) {
+                smash.jobs_list.removeJobById(our_job->job_id);
             }
             smash.pid_in_fg = 0;
         }
@@ -752,7 +760,11 @@ void BackgroundCommand::execute() {
             cerr<<"smash error: bg: job-id " << job_id << " does not exist" << endl;
             return;
         }
-        cout << our_job->command << " : " << our_job->process_id;
+        if (!our_job->is_stopped) {
+            cerr<<"smash error: bg: job-id " << job_id << " is already running in the background" << endl;
+            return;
+        }
+        cout << our_job->command << " : " << our_job->process_id << endl;
         if(kill(our_job->process_id,SIGCONT)==-1){
             perror("smash error: kill failed");
             return;
@@ -842,14 +854,14 @@ RedirectionCommand::RedirectionCommand(const char *cmd_line) : Command(cmd_line)
     }
     int fd;
     if (is_append){
-        fd = open(file_path.c_str(), O_CREAT | O_RDWR | O_APPEND);
+        fd = open(file_path.c_str(), O_CREAT | O_RDWR | O_APPEND, 0777);
         if (fd == -1) {
             perror("smash error: open failed");
             return;
         }
     }
     else {
-        fd = open(file_path.c_str(), O_CREAT | O_RDWR | O_TRUNC);
+        fd = open(file_path.c_str(), O_CREAT | O_RDWR | O_TRUNC, 0777);
         if (fd == -1) {
             perror("smash error: open failed");
             return;
@@ -886,7 +898,7 @@ void RedirectionCommand::execute() {
 
 PipeCommand::PipeCommand(const char* cmd_line): Command(cmd_line){
     string cmd_s = _trim(string(cmd_line));
-    int position =cmd_s.find_last_not_of('|');
+    int position =cmd_s.find('|');
     if(cmd_s[position+1]=='&'){
         second_command_str=cmd_s.substr(position + 2, cmd_s.size());
         stderr_flag=true ;
@@ -977,7 +989,7 @@ void PipeCommand::execute() {
         perror("smash error: close failed");
         return;
     }
-    /*
+
     if(waitpid(pid1, nullptr, WUNTRACED)==-1){
         perror("smash error: waitpid failed");
         return;
@@ -986,7 +998,7 @@ void PipeCommand::execute() {
         perror("smash error: waitpid failed");
         return;
     }
-     */
+
 
 
 }
